+++ app/Controllers/Regulasidata.php
@@ -17,6 +17,13 @@ class Regulasidata extends BaseController
 
     public function index()
     {
+        $isLoggedIn = session('log');
+        $userRole = session()->get('level');
+
+        if (!$isLoggedIn || $userRole !== 'Admin') {
+            // Jika pengguna tidak terautentikasi atau bukan Admin, redirect ke halaman lain
+            return redirect()->to(base_url('home'));
+        }
         $currentpage = $this->request->getVar('page_reg') ? $this->request->getVar('page_reg') : 1;
         $keyword = $this->request->getVar('keyword');
         if ($keyword) {
@@ -37,7 +44,13 @@ class Regulasidata extends BaseController
 
     public function addreg()
     {
+        $isLoggedIn = session('log');
+        $userRole = session()->get('level');
 
+        if (!$isLoggedIn || $userRole !== 'Admin') {
+            // Jika pengguna tidak terautentikasi atau bukan Admin, redirect ke halaman lain
+            return redirect()->to(base_url('home'));
+        }
         $fileinfo = $this->request->getFile('file_reg');
         $fileinfo->move('files/regulasi');
         $namafileinfo = $fileinfo->getName();
@@ -45,7 +58,7 @@ class Regulasidata extends BaseController
         $data = array(
             'judul_reg' => $this->request->getPost('judul_reg'),
             'file_reg' => $namafileinfo,
-            'created_at' => date('Y-m-d H:i:s')            
+            'created_at' => date('Y-m-d H:i:s')
         );
         $this->M_Regulasidata->addreg($data);
         session()->setFlashdata('pesan', 'Data berhasil ditambahkan !!!');
@@ -54,6 +67,13 @@ class Regulasidata extends BaseController
 
     public function editreg($id_reg)
     {
+        $isLoggedIn = session('log');
+        $userRole = session()->get('level');
+    
+        if (!$isLoggedIn || $userRole !== 'Admin') {
+            // Jika pengguna tidak terautentikasi atau bukan Admin, redirect ke halaman lain
+            return redirect()->to(base_url('home'));
+        }
         $data = array(
             'id_reg' => $id_reg,
             'judul_reg' => $this->request->getPost('judul_reg'),
@@ -66,7 +86,13 @@ class Regulasidata extends BaseController
 
     public function delete($id_reg)
     {
-        
+        $isLoggedIn = session('log');
+        $userRole = session()->get('level');
+    
+        if (!$isLoggedIn || $userRole !== 'Admin') {
+            // Jika pengguna tidak terautentikasi atau bukan Admin, redirect ke halaman lain
+            return redirect()->to(base_url('home'));
+        }
         $file_reg = $this->M_Regulasidata->filedata($id_reg);
         unlink('files/regulasi/' . $file_reg['file_reg']);
